+++ datamol/predictors/__init__.py
@@ -1,0 +1,2 @@
+from .esol import esol
+from .esol import esol_from_data
